Split google-play tool implementation into separate action classes
