+++ src/codemagic/tools/google_play/__init__.py
@@ -1,0 +1,1 @@
+from .google_play import GooglePlay
