+++ src/codemagic/tools/google_play/__main__.py
@@ -1,0 +1,3 @@
+if __name__ == '__main__':
+    from .google_play import GooglePlay
+    GooglePlay.invoke_cli()
