+++ src/codemagic/tools/google_play/action_groups/google_play_action_groups.py
@@ -1,0 +1,8 @@
+from codemagic import cli
+
+
+class GooglePlayActionGroups(cli.ActionGroup):
+    TRACKS = cli.ActionGroupProperties(
+        name='tracks',
+        description='Manage your Google Play release tracks',
+    )
