+++ src/codemagic/tools/google_play/actions/__init__.py
@@ -1,0 +1,1 @@
+from .get_latest_build_number_action import GetLatestBuildNumberAction
