+++ src/codemagic/tools/google_play/errors.py
@@ -1,0 +1,5 @@
+from codemagic import cli
+
+
+class GooglePlayError(cli.CliAppException):
+    pass
