+++ tests/tools/test_google_play.py
@@ -8,8 +8,8 @@ from unittest import mock
 import pytest
 
 from codemagic.tools.google_play import GooglePlay
-from codemagic.tools.google_play import GooglePlayArgument
-from codemagic.tools.google_play import Types
+from codemagic.tools.google_play.argument_types import CredentialsArgument
+from codemagic.tools.google_play.arguments import GooglePlayArgument
 
 credentials_argument = GooglePlayArgument.GCLOUD_SERVICE_ACCOUNT_CREDENTIALS
 
@@ -23,7 +23,7 @@ def register_args(cli_argument_group):
 @pytest.fixture()
 def namespace_kwargs():
     ns_kwargs = {
-        credentials_argument.key: Types.CredentialsArgument('{"type":"service_account"}'),
+        credentials_argument.key: CredentialsArgument('{"type":"service_account"}'),
     }
     for arg in GooglePlay.CLASS_ARGUMENTS:
         if not hasattr(arg.type, 'environment_variable_key'):
@@ -47,7 +47,7 @@ def test_missing_credentials_arg(namespace_kwargs):
 
 
 def test_invalid_credentials_from_env(namespace_kwargs):
-    os.environ[Types.CredentialsArgument.environment_variable_key] = 'invalid credentials'
+    os.environ[CredentialsArgument.environment_variable_key] = 'invalid credentials'
     namespace_kwargs[credentials_argument.key] = None
     cli_args = argparse.Namespace(**dict(namespace_kwargs.items()))
     with pytest.raises(argparse.ArgumentError) as exception_info:
@@ -56,7 +56,7 @@ def test_invalid_credentials_from_env(namespace_kwargs):
 
 
 def test_credentials_invalid_path(namespace_kwargs):
-    os.environ[Types.CredentialsArgument.environment_variable_key] = '@file:this-is-not-a-file'
+    os.environ[CredentialsArgument.environment_variable_key] = '@file:this-is-not-a-file'
     namespace_kwargs[credentials_argument.key] = None
     cli_args = argparse.Namespace(**dict(namespace_kwargs.items()))
     with pytest.raises(argparse.ArgumentError) as exception_info:
@@ -64,9 +64,9 @@ def test_credentials_invalid_path(namespace_kwargs):
     assert str(exception_info.value) == 'argument --credentials: File "this-is-not-a-file" does not exist'
 
 
-@mock.patch('codemagic.tools.google_play.GooglePlayDeveloperAPIClient')
+@mock.patch('codemagic.tools.google_play.google_play.GooglePlayDeveloperAPIClient')
 def test_read_private_key(mock_google_play_api_client, namespace_kwargs):
-    namespace_kwargs[credentials_argument.key] = Types.CredentialsArgument('{"type":"service_account"}')
+    namespace_kwargs[credentials_argument.key] = CredentialsArgument('{"type":"service_account"}')
     _ = GooglePlay.from_cli_args(argparse.Namespace(**namespace_kwargs))
     mock_google_play_api_client.assert_called_once_with('{"type":"service_account"}')
 
@@ -76,10 +76,10 @@ def test_read_private_key(mock_google_play_api_client, namespace_kwargs):
         {credentials_argument.type.environment_variable_key: f'@file:{filename}'},
     ),
     lambda filename, ns_kwargs: ns_kwargs.update(
-        {credentials_argument.key: Types.CredentialsArgument(f'@file:{filename}')},
+        {credentials_argument.key: CredentialsArgument(f'@file:{filename}')},
     ),
 ])
-@mock.patch('codemagic.tools.google_play.GooglePlayDeveloperAPIClient')
+@mock.patch('codemagic.tools.google_play.google_play.GooglePlayDeveloperAPIClient')
 def test_private_key_path_arg(mock_google_play_api_client, configure_variable, namespace_kwargs):
     with NamedTemporaryFile(mode='w') as tf:
         tf.write('{"type":"service_account"}')
@@ -96,10 +96,10 @@ def test_private_key_path_arg(mock_google_play_api_client, configure_variable, n
         {credentials_argument.type.environment_variable_key: '@env:CREDENTIALS'},
     ),
     lambda ns_kwargs: ns_kwargs.update(
-        {credentials_argument.key: Types.CredentialsArgument('@env:CREDENTIALS')},
+        {credentials_argument.key: CredentialsArgument('@env:CREDENTIALS')},
     ),
 ])
-@mock.patch('codemagic.tools.google_play.GooglePlayDeveloperAPIClient')
+@mock.patch('codemagic.tools.google_play.google_play.GooglePlayDeveloperAPIClient')
 def test_private_key_env_arg(mock_google_play_api_client, configure_variable, namespace_kwargs):
     os.environ['CREDENTIALS'] = '{"type":"service_account"}'
     namespace_kwargs[credentials_argument.key] = None
